Rename serialize auto derivation object
