+++ macros/src/main/scala/sdk/serialize/SerializeDerivation.scala
@@ -1,4 +1,4 @@
-package sdk.codec
+package sdk.serialize
 
 import cats.Applicative
 import cats.syntax.all.*
