+++ node/src/main/scala/node/Genesis.scala
@@ -5,7 +5,7 @@ import cats.effect.Async
 import cats.syntax.all.*
 import dproc.data.Block
 import node.Hashing.*
-import sdk.codec.SerializeEvalInstances.*
+import sdk.serialize.auto.*
 import sdk.data.{BalancesDeploy, BalancesDeployBody, BalancesState}
 import sdk.diag.Metrics
 import sdk.history.History
