+++ node/src/main/scala/node/api/ExternalApiSlickImpl.scala
@@ -13,7 +13,7 @@ import sdk.codecs.Digest
 import sdk.data.{BalancesDeploy, BalancesDeployBody, BalancesState}
 import sdk.history.ByteArray32
 import sdk.store.KeyValueTypedStore
-import sdk.codec.SerializeEvalInstances.*
+import sdk.serialize.auto.*
 import node.Hashing.*
 import sdk.syntax.all.*
