+++ sdk/src/main/scala/sdk/codecs/Serialize.scala
@@ -8,7 +8,7 @@ package sdk.codecs
  * Any type that is supposed to be transmitted over the wire, or have a digest or stored on chain should have an
  * instance of this type class.
  *
- * Automatic derivation of instances is provided by the [[sdk.codec.SerializeInstances]] in macro project.
+ * Automatic derivation of instances is provided by the [[sdk.serialize.auto]] in macro project.
  * */
 trait Serialize[F[_], A] {
   def write(x: A): PrimitiveWriter[F] => F[Unit]
